# backend/docs/swagger.yaml
basePath: /
definitions:
  internal_httpserver_handlers.ErrorResponse:
    properties:
      error:
        type: string
      ok:
        type: boolean
    type: object
  internal_httpserver_handlers.GenericObject:
    additionalProperties: {}
    type: object
...
  internal_httpserver_handlers.mpPrefsUpdateRequest:
    properties:
      driver_numbers:
        items:
          type: integer
        type: array
      team_keys:
        items:
          type: string
        type: array
      team_name:
        type: string
    type: object
info:
  contact: {}
  description: 后端基于 Gin 的 HTTP API 文档（swaggo/swag 自动生成）。
  title: F1InkDashboard Backend API
  version: 1.0.0
paths:
...
  /api/v1/mp/auth/prefs:
    get:
      description: 返回当前登录用户的偏好（车队/车手关注列表）及附加展示信息。
      parameters:
      - description: 返回格式版本（用于兼容旧版小程序）
        in: query
        name: v
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户偏好
      tags:
      - MiniProgramAuth
    put:
      consumes:
      - application/json
      description: team_keys/driver_numbers 会去重、排序，并最多保留 12 个。
      parameters:
      - description: 返回格式版本（用于兼容旧版小程序）
        in: query
        name: v
        type: string
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpserver_handlers.mpPrefsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新用户偏好
      tags:
      - MiniProgramAuth
...
  /api/v1/mp/news:
    get:
      description: 分页返回新闻条目列表（不包含 content）。
      parameters:
      - default: Asia/Shanghai
        description: IANA 时区名称
        in: query
        name: tz
        type: string
      - default: 1
        description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量（最大 50）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: 小程序新闻列表
      tags:
      - MiniProgram
  /api/v1/mp/news/{id}:
    get:
      description: 返回指定 id 的新闻详情（包含 content）。
      parameters:
      - description: 新闻 ID
        in: path
        name: id
        required: true
        type: string
      - default: Asia/Shanghai
        description: IANA 时区名称
        in: query
        name: tz
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: 小程序新闻详情
      tags:
      - MiniProgram
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Authorization: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
  TokenQuery:
    description: 简单鉴权 token（NEWS_INGEST_TOKEN / OPENF1_INGEST_TOKEN / WECHATPAY_API_TOKEN
      等）
    in: query
    name: token
    type: apiKey
swagger: "2.0"